+++ lib/post/widgets/post_view.dart
@@ -19,6 +19,7 @@ import 'package:thunder/core/singletons/preferences.dart';
 import 'package:thunder/feed/utils/utils.dart';
 import 'package:thunder/feed/view/feed_page.dart';
 import 'package:thunder/post/pages/create_comment_page.dart';
+import 'package:thunder/shared/advanced_share_sheet.dart';
 import 'package:thunder/shared/common_markdown_body.dart';
 import 'package:thunder/thunder/bloc/thunder_bloc.dart';
 import 'package:thunder/core/auth/bloc/auth_bloc.dart';
@@ -50,6 +51,7 @@ class PostSubview extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
+    final bool useAdvancedShareSheet = context.read<ThunderBloc>().state.useAdvancedShareSheet;
 
     final PostView postView = postViewMedia.postView;
     final Post post = postView.post;
@@ -368,20 +370,22 @@ class PostSubview extends StatelessWidget {
                       : null,
                   icon: postView.post.locked
                       ? Icon(Icons.lock, semanticLabel: AppLocalizations.of(context)!.postLocked, color: Colors.red)
-                      : Icon(Icons.reply_rounded, semanticLabel: AppLocalizations.of(context)!.reply),
+                      : Icon(Icons.reply_rounded, semanticLabel: AppLocalizations.of(context)!.reply(0)),
                 ),
               ),
               Expanded(
                 flex: 1,
                 child: IconButton(
                   icon: const Icon(Icons.share_rounded, semanticLabel: 'Share'),
-                  onPressed: postViewMedia.media.isEmpty
-                      ? () => Share.share(post.apId)
-                      : () => showPostActionBottomModalSheet(
-                            context,
-                            postViewMedia,
-                            actionsToInclude: [PostCardAction.sharePost, PostCardAction.shareMedia, PostCardAction.shareLink],
-                          ),
+                  onPressed: useAdvancedShareSheet
+                      ? () => showAdvancedShareSheet(context, postViewMedia)
+                      : postViewMedia.media.isEmpty
+                          ? () => Share.share(post.apId)
+                          : () => showPostActionBottomModalSheet(
+                                context,
+                                postViewMedia,
+                                actionsToInclude: [PostCardAction.sharePost, PostCardAction.shareMedia, PostCardAction.shareLink],
+                              ),
                 ),
               )
             ],