+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Improved account switching logic - contribution from @micahmo
 - Added anonymous browsing mode for any instance - contribution from @micahmo
 - Show trending communities on search page - contribution from @micahmo
+- Added new post sharing options - contribution from @micahmo
+- Added Read All in inbox. - contribution from @ggriffo
 
 ### Changed
 - Collapsed comments are easier to expand - contribution from @micahmo